+++ packages/build-webpack-config/src/setWebpackLoaders.ts
@@ -19,10 +19,13 @@ const configCSSRule = (config, style, loaders = []) => {
     },
   };
   const postcssOpts = {
+    // lock postcss version
+    // eslint-disable-next-line global-require
+    implementation: require('postcss'),
     postcssOptions: {
       config: false,
       // eslint-disable-next-line global-require
-      ...(require('./postcss.config').default()),
+      ...(require('./postcss.config').default()),  
     }
   };