+++ packages/plugin-react-app/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.0.4
+
+- [fix] error occurred when config externals and remoteRuntime
+- [fix] build error when active `remoteRuntime`
+
 ## 2.0.3
 
 - [feat] support apply static loader with query like `?worker|?url|?raw|?worker&inline`